requirements.txt
discord.py
colorlog
setuptools